+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Upgrade HTTP client to okhttp.
+
 ## 1.4.6 (March 16, 2015)
 
 * Fix bug when initializing with user id. Api key was not set properly.